+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/metrics.ts
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 
 import Datasource from '../datasource';
 import { AzureMonitorQuery } from '../types';
-import { convertTimeGrainsToMs } from './common';
+import { convertTimeGrainsToMs } from '../utils/common';
 
 export interface MetricMetadata {
   aggOptions: Array<{ label: string; value: string }>;
