+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/utils/common.ts
