+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/utils/useLastError.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback, useMemo } from 'react';
-import { AzureMonitorErrorish } from '../../types';
+import { AzureMonitorErrorish } from '../types';
 import { messageFromError } from './messageFromError';
 
 type SourcedError = [string, AzureMonitorErrorish];
